pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "legal-contract-risk-reviewer"
version = "0.1.0"
description = "Legal Contract Risk Reviewer - An OpenEnv environment where AI agents analyze contracts for risky terms, missing clauses, and legal loopholes"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "gradio>=4.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["legal_contract_risk_reviewer", "legal_contract_risk_reviewer.server"]
package-dir = { "legal_contract_risk_reviewer" = ".", "legal_contract_risk_reviewer.server" = "server" }
